requirements.txt
asgiref==3.2.3
Django==2.1.15
django-jet==1.0.8
django-pyodbc-azure==2.1.0.0
feedparser==5.2.1
google-api-python-client==1.4.1
httplib2==0.17.0
joblib==0.14.1
numpy==1.18.1
oauth2client==4.1.3
pandas==1.0.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyodbc==4.0.30
python-dateutil==2.8.1
pytz==2019.3
rsa==4.0
scikit-learn==0.22.1
scipy==1.4.1
six==1.14.0
sqlparse==0.3.0
uritemplate==3.0.1
